ImageFilename DO and services moved to Naming sub namespace
